Avoid run ensure block of test on non Windows platform
This was causing RubyGems to fail to remove temporary directory
during test teardown on platforms other than Windows.

+++ test/rubygems/test_gem_compiler.rb
@@ -159,6 +159,8 @@ class TestGemCompiler < Gem::TestCase
 
     assert_includes spec.files, "lib/#{artifact}"
   ensure
+    return unless Gem.win_platform?
+
     FileUtils.rm_rf @tempdir
 
     ENV["TMP"] = old_tmp
